move CameraDeviceWrapper to its own file
